+++ src/stream_proxy.rs
@@ -170,6 +170,7 @@ impl FluxBackgroundTask for StreamProxyTask {
         let _ = shutdown.wait_for_shutdown().await;
         for task in tasks {
             task.abort();
+            let _ = task.await;
         }
     }
 }